3-24-16
--In the Web.Config file in the root directory, in the <appSettings>
block, the “value” string for the “SendGridPassword” key must be filled
in.

--The code is for the online version. To run on the desktop, do the
following.
(1) In the Web.Config file in the root directory, in the
<connectionStrings> block, comment out the “BudgetMaster1” entry and
uncomment the “DefaultConnection” entry.
(2) In the “IdentityModels.cs” file in the “Models” folder, comment out
the “BudgetMaster1” base (with “public ApplicationDbContext” above) and
uncomment the “DefaultConnection” base (with “public
ApplicationDbContext” above).
(3) In the app.js file in NgApps, comment out the BudgetMaster1 URL
serviceBase declaration, and uncomment the localhost one.
(4) In the “SendEmail” method in the “UserController” class in the
“UserController.cs” file, comment out the “budgetmaster1” “string url”
assignment and uncomment the “localhost” one.